+++ NEWS.md
@@ -4,4 +4,12 @@
 
 * **2021-10-31:** Release of first version on Github. Version name: "EpiLPS is born".
 
+### Version 1.0.1 ###
+
+* **2021-12-08:** 
+
+* Version name: "EpiLPS smoothing"
+* The epilps routine can now be fitted to count data containing NAs.
+* Possibility to add dates to the *epilps()* routine.
+